Fix rotation issue when using config_lidOpenRotation
=> only use the config_lidOpenRotation condition if docked.
Because when undocked now the LID is considered "OPEN" and the config_lidOpenRotation is being used at all time if >=0
Only the asus tablets are using that option, so it should be pretty safe to go with that.
It shouldn't change the previous behaviour.

It was working before we integrate the lidpatch into the TF101 kernel,
most likely because the lid was considered "CLOSED" when undocked.
Now with the lidpatch, the logic is inverted compared to before.

config_lidOpenRotation is still needed to have the proper
orientation when docked and prevent the screen from rotating,
so we shouldn't discard that option on devices with docks.

Change-Id: Ia7f9ed79b04dd91440d883c2f53e839258dacb02