Fix crazy draw loop in SystemUI
That was making systemui + surfaceflinger + irq eat about ~20% cpu on the TF101
The problem is that applyGlow and applyKey are called from onDraw
the apply functions also called invalidate in case you don't have a custom color set
generating another OnDraw and so on.

PatchSet2: the animator was reinitializing the filters
=> apply the filters back at the end of the animation.
The animator seems to be eating cpu during the animation too, not sure if it's normal for that kind of animation or not.
We might want to have a look at that later on, and also get the animator to use the custom color scale.

PatchSet3: Add filter calls back into OnDraw but remove invalidate call.
Appears to work just fine and eliminates the wonky draw loop

Change-Id: Ib69213a74ac89a97fd44955faf29226d9798fa75