Add IBSS Ad-Hoc support (3/4)
Credits: http://www.thinktube.com/android-tech/46-android-wifi-ibss

add the necessary bits to the Android framework,
which is mostly passing the necessary configuration (mode and frequency)
to wpa_supplicant and asking wpa_supplicant wether the driver supports IBSS
mode. We are trying to push these API changes thru AOSP.

We have added the following to the public Android API:

WifiConfiguration
    public boolean isIBSS;
    public int frequency;

Which represent the respecive configuration options (PATCH1).

WifiManager
    public boolean isIbssSupported();
    public List<WifiChannel> getSupportedChannels();

Which can be used to ask wether the driver supports IBSS mode (PATCH2) and to get the list of available frequencies (PATCH3).

Note that we can only expect a valid result after WifiManager.WIFI_STATE_ENABLED has been reached.
This is due to the internal state machine handling in WifiStateMachine: we can not expect
wpa_supplicant to be running or drivers loaded at all times.

Change-Id: Id2c515054d6c51b1217fe5ad18e3ea97f3598032