Remove static controllers and cleanly implement EosObserver
Due to reports of possible memory leaks in SystemUI, I've
decided static access to controllers is not the correct
implementation for features. Instead, the observer and
ui controller will be accessible from top level bar service.

I'm not sure if this could be the cause of any leaks,
but looking at the SystemUI structure as a whole, this
feels much more natural and correct.

Change-Id: I4359f13b9c409b8a7ca4c854647224c05f68b1fa